cleanup data visualization code
