--- src/app/utility/visualization/matrix-scatter-plot/matrix-scatter-plot.component.css
@@ -1,11 +1,0 @@
-/* .btn:not(.btn-primary-4):not(.btn-primary-2){
-    border: solid;
-} */
-
-.hidden{
-    display: none;
-}
-
-.btn-primary-2 sup{
-    color: white !important;
-}
