+++ src/app/utility/visualization/visualization.component.css
@@ -1,0 +1,3 @@
+.btn-primary-2 sup{
+    color: white !important;
+}
