+++ src/app/utility/visualization/visualization.component.html
@@ -1,3 +1,29 @@
-<div class="p-2">
-    <app-matrix-scatter-plot></app-matrix-scatter-plot>
+<div class="d-flex bg-white" *ngIf="numberOfSelectedOptions > 1">
+    <div class="d-flex flex-column h-100 w-100">
+        <div class="p-1">
+            <div #matrixPlot id="matrixPlot"></div>
+        </div>
+    </div>
+    <div class="d-flex flex-wrap flex-column pl-2 pr-2 border-left">
+        <div class="btn-group">
+            <button class="btn btn-sm btn-outline" [ngClass]="{'btn-primary-2': dataView == 'splom' && numberOfSelectedOptions > 2}"
+                (click)="setView('splom')" [disabled]="numberOfSelectedOptions < 3">Correlation</button>
+            <button class="btn btn-sm btn-outline" [ngClass]="{'btn-primary-2': dataView == 'heatmap' && numberOfSelectedOptions > 2}"
+                (click)="setView('heatmap')" [disabled]="numberOfSelectedOptions < 3">R<sup>2</sup></button>
+        </div>
+        <div class="pt-1 border-top" *ngFor="let option of meterOptions; let index = index;">
+            <button class="btn w-100 mb-0 btn-sm btn-outline" [ngClass]="{'btn-primary-4': option.selected}"
+                (click)="toggleMeterOption(index)">{{option.meter.name}}</button>
+        </div>
+        <div *ngFor="let option of predictorOptions; let index = index;">
+            <button class="btn w-100 mb-0 btn-sm btn-outline" [ngClass]="{'btn-primary-4': option.selected}"
+                (click)="togglePredictorOption(index)">{{option.predictor.name}}</button>
+        </div>
+    </div>
+</div>
+
+<div class="d-flex p-4 bg-white" *ngIf="numberOfSelectedOptions < 2">
+    <p class="alert alert-warning text-center">
+        You have one or fewer Meters and Predictors. More then one is needed to view data correlations.
+    </p>
 </div>
